RestFilm/requirements.txt
certifi==2018.11.29
chardet==3.0.4
defusedxml==0.5.0
Django==2.1.4
django-allauth==0.38.0
django-filebrowser-no-grappelli==3.7.5
django-modeltranslation==0.12.1
django-robots==3.1.0
django-tinymce4-lite==1.7.3
idna==2.7
jsmin==2.2.2
oauthlib==2.1.0
Pillow==5.3.0
pyenchant==2.0.0
python3-openid==3.1.0
pytz==2018.7
requests==2.20.1
requests-oauthlib==1.0.0
urllib3==1.24.1
